No longer increment anonymous users
The achievements service no longer handles user names, and so it doesn't
disambiguate between anonymous users.  We'd rather not track anonymous
user statistics at all, since they can't earn wiki points or
achievements.  The only purpose of keeping those stats was for site
point weights calculation (since that is based on the number of edits
performed recently) and for Special:Analytics, which will likely come
from a different data source in the future.