Defer all stat increments to a job
Cheevos currently attempts to perform all stat increments in a shutdown
function, enqueueing a job if this fails. This can result in excessive
load on the backing service as some stats are updated on e.g. every
registered user PV on GP wikis, which can in turn also overload the UCP
process pool due to the synchronous update.

So, always defer these stat increments to a job instead, for better
concurrency control and reduced impact on user-facing process pools.